Further cleanup, some basic documentation and re-organization.
