+++ src/main/scala/com/nerdery/soundchallenge/services/midigenerators/dsl/DslMidiGenerator.scala
@@ -2,25 +2,24 @@ package com.nerdery.soundchallenge.services.midigenerators.dsl
 
 import javax.sound.midi.{Sequence, Track}
 
-import com.nerdery.soundchallenge.services.midigenerators.FreeformMidiGenerator
-
 /**
  *
  * @author Justin Mullin
  *
- * Exposes a little Scala DSL for MIDI composition.
+ * Exposes a little Scala DSL for MIDI composition.  See SuperMarioMidiGenerator for an
+ * example of the DSL in action.
  *
  */
 trait DslMidiGenerator extends FreeformMidiGenerator {
   implicit def intToKey(i: Int) = Key(i, 0)
 
   override def getName: String
+
   override def build(sequence: Sequence): Unit = {
-    implicit val seq = sequence
-    generate()(seq)
+    generate(sequence)
   }
 
-  def generate()(implicit sequence: Sequence): Unit
+  def generate(implicit sequence: Sequence): Unit
 
   def track(name: String, voice: Int, tempo: Int, channel: Int=0)(build: Track => Run)(implicit sequence: Sequence): Unit = {
     implicit val newTrack = buildTrack(sequence, name, tempo)
@@ -39,7 +38,6 @@ trait DslMidiGenerator extends FreeformMidiGenerator {
   }
 
   def add(tick: Long, note: Note, channel: Int)(implicit track: Track): Unit = {
-    println(f"Adding note ${note.key.value} at $tick for ${note.duration.ticks}")
     addNote(track, tick, note.key.value, channel, 0x70, note.duration.ticks)
   }
 }
