+++ src/main/scala/com/nerdery/soundchallenge/services/midigenerators/dsl/Interval.scala
@@ -1,5 +1,8 @@
 package com.nerdery.soundchallenge.services.midigenerators.dsl
 
+/**
+ * Represents a musical interval between two keys
+ */
 case class Interval(halfSteps: Int)
 
 object Interval {
