+++ src/main/scala/com/nerdery/soundchallenge/services/midigenerators/dsl/Note.scala
@@ -1,5 +1,8 @@
 package com.nerdery.soundchallenge.services.midigenerators.dsl
 
+/**
+ * Represents a key played with a certain duration
+ */
 case class Note(key: Key, duration: Duration) extends Run {
   override def expand(start: Long) = Seq(NoteEvent(start, this))
   override def length = duration.ticks
