+++ src/main/scala/com/nerdery/soundchallenge/services/midigenerators/dsl/NoteEvent.scala
@@ -1,0 +1,6 @@
+package com.nerdery.soundchallenge.services.midigenerators.dsl
+
+/**
+ * Represents a note played at a certain time
+ */
+case class NoteEvent(tick: Long, note: Note)
